+++ app/src/main/java/com/sforge/quotes/activity/CreateQuotes.java
@@ -4,48 +4,20 @@ import android.content.res.Configuration;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.EditText;
-import android.widget.Toast;
-
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.res.ResourcesCompat;
-
-import com.google.firebase.auth.FirebaseAuth;
 import com.sforge.quotes.R;
-import com.sforge.quotes.entity.Quote;
-import com.sforge.quotes.repository.QuoteRepository;
-import com.sforge.quotes.repository.UserQuoteRepository;
 
 import java.util.Objects;
 
 public class CreateQuotes extends AppCompatActivity {
-
-    private String user;
-    private final int quoteLengthLimit = 200;
-    private final int authorLengthLimit = 50;
-    private final int minQuoteLengthLimit = 5;
-    private final int minAuthorLengthLimit = 2;
-
-    private QuoteRepository quoteRepository;
-    private EditText createQuoteEditText, createAuthorEditText;
-    private UserQuoteRepository userQuoteRepository;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_create_quotes);
 
-        createActionBar();
 
-        createQuoteEditText = findViewById(R.id.createQuoteEditText);
-        createAuthorEditText = findViewById(R.id.createAuthorEditText);
-
-        quoteRepository = new QuoteRepository();
-
-        user = Objects.requireNonNull(FirebaseAuth.getInstance().getCurrentUser()).getUid();
-
-        userQuoteRepository = new UserQuoteRepository(user);
     }
 
     @Override
@@ -58,35 +30,10 @@ public class CreateQuotes extends AppCompatActivity {
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         if (item.getItemId() == R.id.create_quotes_menu_done) {
-            if (createQuoteEditText.getText().toString().length() <= quoteLengthLimit && createAuthorEditText.getText().toString().length() <= authorLengthLimit
-                    && createQuoteEditText.getText().toString().length() >= minQuoteLengthLimit && createAuthorEditText.getText().toString().length() >= minAuthorLengthLimit) {
-                createQuote();
-                finish();
-            } else if (createQuoteEditText.getText().toString().length() > quoteLengthLimit) {
-                createQuoteEditText.setError("Quote Length Limit is " + quoteLengthLimit + " Letters!");
-                createQuoteEditText.requestFocus();
-            } else if (createAuthorEditText.getText().toString().length() > authorLengthLimit) {
-                createAuthorEditText.setError("Author Length Limit is " + authorLengthLimit + " Letters!");
-                createAuthorEditText.requestFocus();
-            } else if (createQuoteEditText.getText().toString().length() < minQuoteLengthLimit) {
-                createQuoteEditText.setError("Quote Must Have at Least " + minQuoteLengthLimit + " Letters!");
-                createQuoteEditText.requestFocus();
-            } else if (createAuthorEditText.getText().toString().length() < minAuthorLengthLimit) {
-                createAuthorEditText.setError("Author's Name Must Have At Least " + minAuthorLengthLimit + " Letters!");
-                createAuthorEditText.requestFocus();
-            }
+
         }
 
         return super.onOptionsItemSelected(item);
-    }
-
-    public void createQuote(){
-        Quote quote = new Quote(createQuoteEditText.getText().toString(), createAuthorEditText.getText().toString(), user);
-        String key = userQuoteRepository.addWithKeyReturn(quote);
-        quoteRepository.addWithKey(key, quote)
-                .addOnSuccessListener(suc -> Toast.makeText(this, "Quote is Added Successfully!", Toast.LENGTH_SHORT).show())
-                .addOnFailureListener(er-> Toast.makeText(this, "Failed To Add the Quote", Toast.LENGTH_SHORT).show())
-                .addOnCanceledListener(() -> Toast.makeText(this, "Canceled", Toast.LENGTH_SHORT).show());
     }
 
     public void createActionBar(){
