+++ app/src/main/java/com/sforge/quotes/adapter/CollectionActivityAdapter.java
@@ -4,22 +4,26 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
 import androidx.annotation.NonNull;
-
 import com.firebase.ui.database.FirebaseRecyclerAdapter;
 import com.firebase.ui.database.FirebaseRecyclerOptions;
 import com.google.firebase.database.DataSnapshot;
 import com.sforge.quotes.R;
-import com.sforge.quotes.activity.CollectionsActivity;
 import com.sforge.quotes.view.CollectionActivityVH;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class CollectionActivityAdapter extends FirebaseRecyclerAdapter<DataSnapshot, CollectionActivityVH> {
+
+    public interface OnItemClickListener {
+        void onItemClick(String itemName);
+    }
+
     Context context;
     List<DataSnapshot> list = new ArrayList<>();
+
+    private OnItemClickListener listener;
     /**
      * Initialize an adapter that listens to a Firebase query. See
      * {@link FirebaseRecyclerOptions} for configuration options.
@@ -29,6 +33,10 @@ public class CollectionActivityAdapter extends FirebaseRecyclerAdapter<DataSnaps
     public CollectionActivityAdapter(Context context ,@NonNull FirebaseRecyclerOptions<DataSnapshot> options) {
         super(options);
         this.context = context;
+    }
+
+    public void setOnItemClickListener(OnItemClickListener listener) {
+        this.listener = listener;
     }
 
     public List<DataSnapshot> getList() {
@@ -43,7 +51,9 @@ public class CollectionActivityAdapter extends FirebaseRecyclerAdapter<DataSnaps
         holder.itemView.setOnClickListener(view -> {
             String itemName = holder.name.getText().toString();
 
-            ((CollectionsActivity) context).onClickCalled(itemName);
+            if (listener != null) {
+                listener.onItemClick(itemName);
+            }
         });
     }
