+++ app/src/main/res/values-night/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.Quotes" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.Quotes" parent="Theme.Material3.Light.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/black</item>
         <item name="colorPrimaryVariant">@color/grey</item>
@@ -10,10 +10,10 @@
         <item name="colorSecondaryVariant">@color/black</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">@color/color_primary_variant</item>
+        <item name="android:statusBarColor">@color/color_primary_variant</item>
         <!-- Customize your theme here. -->
     </style>
-    <style name="Theme.AppCompat.Quotes.NoActionBar" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+    <style name="Theme.AppCompat.Quotes.NoActionBar" parent="Theme.Material3.Dark.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/black</item>
         <item name="colorPrimaryVariant">@color/black</item>
@@ -23,7 +23,7 @@
         <item name="colorSecondaryVariant">@color/black</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">@color/color_primary_variant</item>
+        <item name="android:statusBarColor">@color/color_primary_variant</item>
         <!-- Customize your theme here. -->
         <item name="alertDialogTheme">@style/AlertDialogTheme</item>
     </style>
@@ -38,12 +38,12 @@
         <item name="backgroundTint">@color/white</item>
     </style>
     <style name="addButton">
-        <item name="android:background">@drawable/button_round_corners</item>
         <item name="backgroundTint">@color/dark_grey</item>
+        <item name="android:background" tools:targetApi="34">?attr/colorPrimary</item>
         <item name="android:drawableStart">@drawable/ic_baseline_add_24</item>
         <item name="drawableStartCompat">@drawable/ic_baseline_add_24</item>
     </style>
-    <style name="transparentButton" parent="@style/Widget.AppCompat.Button.Borderless">
+    <style name="transparentButton" parent="@style/Widget.Material3.Button">
         <item name="android:background">@drawable/button_round_corners</item>
         <item name="backgroundTint">@color/transparent</item>
         <item name="elevation">0dp</item>
@@ -57,7 +57,7 @@
     </style>
 
 
-    <style name="Theme.AppCompat.SlidrActivityTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+    <style name="Theme.AppCompat.SlidrActivityTheme" parent="Theme.Material3.Dark.NoActionBar">
         <item name="android:windowIsTranslucent">true</item>
         <item name="android:windowBackground">@android:color/transparent</item>
         <item name="android:background">@color/dark_grey</item>
@@ -70,23 +70,14 @@
         <item name="colorSecondaryVariant">@color/black</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">@color/color_primary_variant</item>
+        <item name="android:statusBarColor">@color/color_primary_variant</item>
         <item name="alertDialogTheme">@style/AlertDialogTheme</item>
     </style>
 
     <style name="Theme">Theme</style>
 
-    <style name="AlertDialogTheme" parent="ThemeOverlay.AppCompat.Dialog.Alert">
-        <item name="buttonBarNegativeButtonStyle">@style/NegativeButtonStyle</item>
-        <item name="buttonBarPositiveButtonStyle">@style/PositiveButtonStyle</item>
-    </style>
+    <style name="AlertDialogTheme" parent="ThemeOverlay.Material3.Dialog.Alert">
 
-    <style name="NegativeButtonStyle" parent="Widget.AppCompat.Button.ButtonBar.AlertDialog">
-        <item name="android:textColor">@color/white</item>
-    </style>
-
-    <style name="PositiveButtonStyle" parent="Widget.AppCompat.Button.ButtonBar.AlertDialog">
-        <item name="android:textColor">@color/white</item>
     </style>
 
 </resources>
